+++ core/core-common/src/models/contact.rs
@@ -4,7 +4,9 @@ use std::iter;
 use std::time::Instant;
 
 use crate::actions::contacts::Delete as ContactsDelete;
-use crate::datatypes::{GroupedContacts, LabelType, Labels, LocalContactId};
+use crate::datatypes::{
+    ContactSuggestion, DeviceContact, GroupedContacts, LabelType, Labels, LocalContactId,
+};
 use crate::models::{ContactCard, ContactEmail, ModelExtension, ModelIdExtension};
 use crate::{ContactError, CoreContextError, CoreContextResult};
 use futures::future::try_join;
@@ -415,6 +417,42 @@ impl Contact {
         Ok(GroupedContacts::from_contacts_and_groups(
             contacts,
             contact_groups,
+        ))
+    }
+
+    /// Returns a list of contact suggestions (used for example in Composer). Sorted, deduplicated and filtered by the query.
+    ///
+    /// # Parameters
+    ///
+    /// * `query` - a plaintext string provided by the user that we need to complete
+    /// * `device_contacts` - contacts stored in the device storage, not shared between proton clients.
+    /// * `tether` - The database interface
+    ///
+    /// # Errors
+    ///
+    /// when querying the database fails.
+    ///
+    pub async fn contact_suggestions(
+        query: &str,
+        device_contacts: Vec<DeviceContact>,
+        tether: &Tether,
+    ) -> Result<Vec<ContactSuggestion>, StashError> {
+        // TODO (ET-1971): Extend that implementation by using query to filter contacts, groups and device contacts
+        let (_query,) = (query,);
+
+        let (mut contacts, contact_groups) = try_join!(
+            Contact::find("WHERE deleted = 0", vec![], tether),
+            Label::find_by_kind(LabelType::ContactGroup, tether)
+        )?;
+
+        for contact in &mut contacts {
+            contact.emails(tether).await?;
+        }
+
+        Ok(ContactSuggestion::from_contacts_and_device_contacts(
+            contacts,
+            contact_groups,
+            device_contacts,
         ))
     }