+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-01-XX
 
+## [0.38.0] - 2025-01-28
+
 ### Added
 
   - [ET-1954] New method on `MailUserSession` `connection_status` and `ConnectionStatus` enum.